requirements_draft.txt
python=3.9
numpy=1.26.4
energyflow=1.3.2
pandas=2.2.2 #not used until now
matplotlib=3.8.4 #not used until now
awkward=2.6.3
vector=1.4.0
scikit_learn=1.4.2
particle=0.24.0
pytorch=2.3.0
torchinfo=1.8.0
torch_geometric=2.4.0
extras_torch_geometric
pennylane=0.36.0